+++ packages/ui/src/__generated/ve/components/Menu/styles.css.js
@@ -2,7 +2,7 @@
 import { createRuntimeFn as _7a468 } from "@vanilla-extract/recipes/createRuntimeFn";
 var menuDivider = "_14bi7ra5";
 var menuHeading = "_14bi7ra6";
-var menuItemVariants = _7a468({ defaultClassName: "_14bi7ra2 mt0ih24e mt0ih24y mt0ih237 mt0ih23r mt0ih21x mt0ih2sz", variantClassNames: { selected: { true: "_14bi7ra3", false: "_14bi7ra4" } }, defaultVariants: { selected: false }, compoundVariants: [] });
+var menuItemVariants = _7a468({ defaultClassName: "_14bi7ra2 mt0ih24e mt0ih24y mt0ih237 mt0ih23r mt0ih2sz mt0ih21x mt0ih2tn", variantClassNames: { selected: { true: "_14bi7ra3", false: "_14bi7ra4" } }, defaultVariants: { selected: false }, compoundVariants: [] });
 var menuList = "_14bi7ra0";
 export {
   menuDivider,