+++ sdk/client/src/listeners/network-listener/utils/utils.ts
@@ -1,6 +1,7 @@
 import { RequestResponsePair } from './models'
 import { sanitizeResource } from './network-sanitizer'
 import { getSessionSecureID } from '../../../utils/sessionStorage/highlightSession'
+import { getActiveSpan } from '../../../otel'
 
 export const HIGHLIGHT_REQUEST_HEADER = 'X-Highlight-Request'
 
@@ -299,9 +300,17 @@ function makeId(length: number) {
 	return result
 }
 
-export const createNetworkRequestId = () => {
+export const createNetworkRequestId = (useOtelTraceId?: boolean) => {
 	// Long enough to avoid collisions, not long enough to be unguessable
-	return [getSessionSecureID(), makeId(10)]
+	const requestId = makeId(10)
+
+	if (useOtelTraceId) {
+		const context = getActiveSpan()
+		const traceId = context?.spanContext().traceId
+		return [getSessionSecureID(), traceId ?? requestId]
+	}
+
+	return [getSessionSecureID(), requestId]
 }
 
 export const getHighlightRequestHeader = (